Convert Cucumber step files to TypeScript for better type safety
- Convert all step files from .cjs to .ts
- Add proper TypeScript types and imports
- Add ts-node and @types/chai dependencies
- Update cucumber configuration to use TypeScript
- Add missing step definition for 'the lesson loads successfully'
- All 6 scenarios and 35 steps now pass with TypeScript
