+++ frontend/package.json
@@ -45,6 +45,7 @@
     "@testing-library/jest-dom": "^6.1.0",
     "@testing-library/react": "^14.0.0",
     "@testing-library/user-event": "^14.5.0",
+    "@types/chai": "^5.2.2",
     "@types/jest": "^29.5.14",
     "chai": "^5.2.1",
     "concurrently": "^9.2.0",
@@ -56,7 +57,8 @@
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
     "playwright": "^1.38.0",
-    "prettier": "^3.0.0"
+    "prettier": "^3.0.0",
+    "ts-node": "^10.9.2"
   },
   "scripts": {
     "start": "react-scripts start",
